Further separating out getting files to `browse.services.anypath`
Moves `APath` type. Fixes some comments.

ARXIVCE-289
