+++ browse/services/prevnext/FSPrevNext.py
@@ -2,12 +2,10 @@
 
 from typing import Optional, List
 
-from browse.services.anypath import to_anypath
+from browse.services.anypath import to_anypath, fs_check
 
 from browse.domain.identifier import Identifier, IdentifierException
 from browse.services.documents.fs_implementation.legacy_fs_paths import FSDocMetaPaths
-
-from browse.services import fs_check
 
 from .prevnext_base import PrevNextService, PrevNextResult
